requirements.txt
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-3-jet==1.0.8
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
google-api-core==1.17.0
google-api-python-client==1.8.0
google-auth==1.14.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.2
idna==2.9
inflection==0.4.0
itypes==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.3
Pillow==7.1.1
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.3
pytz==2019.3
requests==2.23.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
uritemplate==3.0.1
urllib3==1.25.9
